[callgraph] Centralize the calling of destructors in the CallGraph of CallGraphState.
Previously this was scattered in various places . Now we have a simple model that
all CallGraphEdges, CallGraphNodes, and CallGraphSCCs are owned by the CallGraph
since they are allocated by the CallGraph's BumpPtrAllocator.

This additionally fixes a problem where we were calling delete on SCC's owned by
the BumpPtrAllocator.

Swift SVN r25522