[+0 self] Behind a flag, add support to the runtime for tracking leaks and emitting structured JSON output of the leaks.
This is just meant to be used in unit test like situations where we want
to only detect leaks in a specific section of code and more importantly
want output that is easily scriptable.

The actual impact on the runtime itself is very small, 2x macro calls that are
compiled out when the given flag is not set. After +0 self, this will be ripped
out.

rdar://19933044

Swift SVN r25537