stdlib: add an experimental serialization API
The primary motivation for doing this is that NSCoding does not work
with struts and Swift containers.  Making everything classes just to
satisfy NSCoding forces unnatural code.

This API requires two times (!) less boilerplate than NSCoding, since
the same method is used for serialization and deserialization.  This API
is also more type-safe, it does not require the user to write 'as' type
casts, unlike NSCoding.

Please take a look at validation-test/stdlib/Serialization.swift to see
the intended use pattern.

The performance of the underlying implementation is already decent, and
there's a lot of room for improvement.

These APIs will be used for writing automation tools in Swift.  Just
like SwiftExperimental, this module is not exposed to extrenal users.

Swift SVN r25678