Verifier: The alloc_stack stack does not need to be consistent on purely unreachable paths.
When we clone loops with exit blocks that are unreachable the verifier
complained before this patch.

loop:
  = alloc_stack
  cond_br %overflow_cond, loop_latch, unreachable_only_exit

loop_latch:
  cond_br  loop, exit

unreachable_only_exit:
  ..
  unreachable

We don't care about stack deallocation on paths that must lead to
unreachable.

The alternative to weaken the verifier for cases like this would be to clone
loop exits that are guaranteed to lead to unreachable which would be a waste of
space.

Swift SVN r25732