[ClangImporter] Handle Clang diagnostics purportedly on line 0.
We don't handle a line 0 in Swift, and this led to LLDB crashing. I'm not
actually sure why Clang was emitting an error on line 0, either; the full
location was "<built-in>:0:4" and the error was

  module 'CoreFoundation' found in both '.../CoreFoundation-17XE9QM744UO5.pcm'
  and '.../CoreFoundation-2BVMOFFOYI8X.pcm'

which turned out to be another manifestation of rdar://problem/19817341
(this was an older LLDB build). No test cases because I'm not sure how to
trigger this, especially not since the previous Radar was fixed, but I'll
clone it to track down why Clang is using such an unusual location.

rdar://problem/19951280

Swift SVN r25766