Fix the last part of <rdar://19981118> Swift 1.2 beta 2: Closures nested in @noescape closures copy, rather than reference, captured vars.
This handles merging flag sets when an inner closure and outer closure both
close over a decl, and have different escaping behaviors.  The outer closure
needs to choose the more conservative capture model (it can't capture just the
address of the decl if the inner closure needs a box).



Swift SVN r25773