Reduce the number of conversions between String and UTF16 when performing string to int conversions.
Those conversions are supposed to be very cheap, but they don't get optimized away yet. Therefore, perform the String to utf16 conversion only once and then reuse it.

This change together with the previous patch which gets rid of a closure inside UTF16View brings the performance of the regressed StrToInt benchmark back to normal.

Swift SVN r25780