Add PrettyStackTrace support for SILFunctionTransforms and SILModuleTransforms.
In the case of a crash in the optimizer or verifier, this prints out information
about the source of the crash. This saves engineering time by allowing one to
get some quick information about the crash without needing to jump into the
buffer.

In the case of a SILFunctionTransform, this prints out the name of the function
being optimized and the name of the transform.

In the case of a SILModuleTransform, this prints out the name of the transform.

<rdar://problem/19946491>

Swift SVN r25831