[+0 self] Add a cmake flag to enable guaranteed self.
The reason why this is necessary is that I need to update parts of the runtime
for +0 self. I am going to use a -D flag to do this implying I need the build
system's help. Since I already have the flag I am also going to wire up the
emission of +0 self code from the compiler to be triggered off this as well.

It is of course disabled by default.

<rdar://problem/20080934>

Swift SVN r25833