Remove a source of potential non-determinism in call graph construction.
Sort call graph nodes in the CalleeSet (which is a SmallPtrSet) by the
ordinal we've assigned before iterating over them.

The caller edge list will change to a SmallPtrSet soon as well, and
technically this won't be necessary at that point, but I will probably
leave the sorting in as it doesn't hurt anything and will ensure that if
we need to change data structures again we'll continue to be
deterministic.

Swift SVN r26091