[sil-verify] Remove the requirement that "from" and "to" types of a cast should not be the same.
We discussed it with Joe and this requirement does not make sense, especially if casts are produced by one pass, but optimized by another one. Moreover, having equal from and to types is eventually not very efficient, but is not semantically wrong. Such casts can often be produced by the inliner or specialised.

Swift SVN r26110