Reimplement emission of StmtCondition (e.g. if/let & while/let) to be in terms of the pattern binding/emission facilities that are currently used for switches.  They are more general (handling all patterns, not hacked up just for optionals).
This leads to us producing better code for if/let bindings, because we 
don't alloc_stack a temporary and deal with memory for non-address-only
types (e.g. the common case of an optional pointer).  Instead, the code 
emits a select_enum{_addr} on the value.

While this changes the generated code in the compiler, there is no exposed
behavioral change to the developer.




Swift SVN r26142