Get TypeChecker::conformsToProtocol() out of the business of checking the conformances.
conformsToProtocol() was doing a confusing double-duty of querying
whether a type conforms to a protocol and initiating a complete check
of the witnesses in a conformance. Split it out: conformsToProtocol()
handles the query, and resolveConformance() does the complete check of
a conformance.

This is still messier than the end result should be. More baby steps.

Swift SVN r26148