Fix crash during devirtualization.
This was exposed by a previous change that fixed two other issues, but
introduced this.

I've fixed the crash (rdar://problem/20165943) and added a test case for
it, but one result of the fix is that we'll currently miss a couple of
cases we could devirtualize, and will instead speculatively devirtualize
in these cases.

I've opened a radar so that I can go back and reenable these cases after
further refactoring that will make it simpler to handle these correctly
(rdar://problem/20166171).

Swift SVN r26152