Temporarily disable configuration of watchOS builds
The SDK isn't available yet, so don't try to configure it. Once it is
available, not everyone may be able to install it, so we will need to do
something more intelligent and configure all but only the SDKs that are
installed into the active Xcode bundle.

Swift SVN r26331